test: Fix bugs, add README, improve makefile
