+++ golang/cosmos/agoricinterchaintest/util.go
@@ -1,18 +1,14 @@
 package agoricinterchaintest
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"strings"
 	"testing"
-	"time"
 
 	interchaintest "github.com/agoric-labs/interchaintest/v6"
-	"github.com/agoric-labs/interchaintest/v6/conformance"
 	"github.com/agoric-labs/interchaintest/v6/ibc"
 	"github.com/agoric-labs/interchaintest/v6/relayer"
-	"github.com/agoric-labs/interchaintest/v6/testreporter"
 	"go.uber.org/zap/zaptest"
 )
 
@@ -201,51 +197,4 @@ func getRelayerFactory(t *testing.T) interchaintest.RelayerFactory {
 	t.Logf("RelayerNmae: %s[%s]", ENV_RELAYERNAME, relayerName)
 
 	return ret
-}
-
-func TestConformance(t *testing.T) {
-	cs := getChainSpec(t)
-	rf := getRelayerFactory(t)
-	testConformance(t, cs[0:2], rf)
-}
-
-func testConformance(t *testing.T, cs []*interchaintest.ChainSpec, rf interchaintest.RelayerFactory) {
-	cf := interchaintest.NewBuiltinChainFactory(zaptest.NewLogger(t), cs)
-
-	ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(45*time.Minute))
-	defer cancel()
-
-	// For our example we will use a No-op reporter that does not actually collect any test reports.
-	rep := testreporter.NewNopReporter()
-
-	// Test will now run the conformance test suite against both of our chains, ensuring that they both have basic
-	// IBC capabilities properly implemented and work with both the Go relayer and Hermes.
-	conformance.Test(t, ctx, []interchaintest.ChainFactory{cf}, []interchaintest.RelayerFactory{rf}, rep)
-}
-
-func TestChainPair(t *testing.T) {
-	cs := getChainSpec(t)
-	rf := getRelayerFactory(t)
-	testChainPair(t, cs[0:2], rf)
-}
-
-func testChainPair(t *testing.T, cs []*interchaintest.ChainSpec, rf interchaintest.RelayerFactory) {
-
-	cf := interchaintest.NewBuiltinChainFactory(zaptest.NewLogger(t), cs)
-
-	// Use a 45 min
-	// ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(45*time.Minute))
-	// defer cancel()
-	ctx := context.Background()
-
-	// For our example we will use a No-op reporter that does not actually collect any test reports.
-	rep := testreporter.NewNopReporter()
-
-	chains, err := cf.Chains(t.Name())
-	if err != nil {
-		panic(fmt.Errorf("failed to get chains: %v", err))
-	}
-
-	client, network := interchaintest.DockerSetup(t)
-	conformance.TestChainPair(t, ctx, client, network, chains[0], chains[1], rf, rep, nil)
 }
